NativeFileSystem was renamed to FileSystemAccess.
Also, NativeFileSystem origin trial was removed.

Chromium changes:

https://source.chromium.org/chromium/chromium/src/+/07ecb97fd6498002303f162d9e6cc5f248599eff

commit 07ecb97fd6498002303f162d9e6cc5f248599eff
Author: Austin Sullivan <asully@chromium.org>
Date:   Fri Jan 8 15:16:17 2021 +0000

    Rename native_file_system -> file_system_access part 3/x.

    This finishes renaming in blink, with the exception of web platform
    tests.

    Renames "NativeFileSystem" to FileSystemAccess or FileSystem.
    Generally, NativeFileSystem is renamed to FileSystemAccess.
    In cases where IDL exposes something as "FileSystem", NativeFileSystem
    is renamed to "FileSystem".

    Bug: 1110509

https://source.chromium.org/chromium/chromium/src/+/9926254ea8d51ce3845d468262eb5cfa5020754f

commit 9926254ea8d51ce3845d468262eb5cfa5020754f
Author: Austin Sullivan <asully@chromium.org>
Date:   Fri Jan 8 16:48:02 2021 +0000

    [FSA] Remove NativeFileSystem2 origin trial

    The origin trial is no longer needed, since the feature has already
    shipped.

    This change was originally proposed in this refactor [1], but was split
    into another CL.

    [1] https://crrev.com/c/2601576

    Bug: 1110509
